Dialer: The speed dial show "not set" after change the name
In the program, it use the name to match the contact and the speed
dial show "not set" when name can't match contact in database, it
can't match the contact after user change the name, so the speed show
"not set".

Use the number to match the contact in database.

CRs-Fixed: 526000

Change-Id: I437bdd3744a9f0368b941077a91d58ca743626ec