Dialer:Fix the speed dial number issue.
Because only the contact name and number are stored in the
sharedpreference,when the contact name is modified in the
contact database,the contact name stored in the sharedpreference
can not be modified correspondly.So when go to speed dial
the tip will be pop up again.

Store the contact ID to sharedpreference,and get the ID after we
long click a number in dialer to choose a contact as the speed dial
contact.

CRs-Fixed: 526035

Change-Id: I0761f35a7d1048084b39f5c0953338efa3f85905