Dialer: Fix the sim speed dial disappeared after turn off airplane mode
When we turn on airplane mode, the sim contact will be deletd from
the contacts database. Then turn off, we enter the speed dial setting,
it will check each speed dial number is exit in the database or not,
if not exit, it would be deleted from SharedPreferences, so the sim
contact would not show in the speed dial list.

 - Add a message telling the user to close the Airplane mode when
   use try to use the speed dial in airplane mode.
 - When enter the speed dial setting and check the number is not
   exit in the database, if it is exit in database, show the name
   in the list item else show the number in the list item.

CRs-Fixed: 509128

Change-Id: I518c7ece7971392597a0be0af6876d39e60e666f