Dialer: A blank view with slot information will display firstly
When we tap "Return to call in progress", returnToInCallScreen() in
DialpadFragment.java will be called. This function will show
InCallScreen and finish the DialtactsActivity by calling
getActivity().finish(), then the SmartDialpadFragment's onPause()
is called and onPause() will finally call hideDialPadShowList(false)
which shows the mDialButtonContainer to UI. However, onDestroy() will
be called at last and during this interval, the "blank view with slot
information" will display.

We should return from onPause() when we detect the Activity is
finishing to avoid UI operations.

CRs-Fixed: 549883

Change-Id: Ic10abe3666c98c6281cc3f69059ce13391bd28dd