Dialer: Do not show DialButton when DialpadButton is visible
In setDialButtonVisibility method, the DialButton will set visible if the
DialpadButton is not showing. But when the DialtactsActivity is not in the
foreground, the DialpadButton is not showing even though it is visible.
It cause a result that DialButton is set as visible too, then these two
buttons came out together.

Use the DialpadButton is visible or not to control the DialButton,
and set DialButton visibility gone if DialpadButton is visible.

CRs-Fixed: 681887

Change-Id: Ia73da43ab91b6ece74721042701c0cefffad54b5