Dialer: Do not let call button cover history button
When the mDialButtonSub1 and mDialButtonSub2 are not visible at
the same time, the width of the mDialButtonSub1 will be "match_parent",
then the call button will cover history button.

Draw the history button after call button, to make sure
the history button will not be covered by cal button.

CRs-Fixed: 689066

Change-Id: Ibf7b321cf41efe6121cb3204339b5df18712b87d