Dialer: Don't call finish when calling from the dialpad
In DSDS or DSDA mode, if there is at least one sim card in phone,
MO a call through the call button in DialpadFragment,
it will call getActivity().finish() to finish the DialtactsActivity,
so when end the call, it will goes to home screen.

Don't call finish when calling from the dialpad.

CRs-Fixed: 685376

Change-Id: I90b06e6594cb479fa49e1e2922a87f67c9f74dd5