Dialer: Do not let call button cover history button
When the mDialButtonSub1 and mDialButtonSub2 are not visible at
the same time, the width of the mDialButtonSub1 will be "match_parent",
then the call button will cover history button.

Draw the history button after call button, to make sure
the history button will not be covered by cal button.

CRs-Fixed: 773012

Change-Id: I0eda4318d87ed82bb5a1a7cf605fa4c948222bf0