Dialer: Fix missed call is cleared even in background
When there is a new missed call, PhoneFavoriteFragment
will call fetchCalls(), and in onCallsFetched(), call
clearMissedCallNotification(). When PhoneFavoriteFragment is in
background, isVisible() return true, so it will clear
missed call notification.

Use isResumed() instead of isVisible(). when PhoneFavoriteFragment
is in background, isResumed() will return false, and it will
return true, when PhoneFavoriteFragment is in foreground.

CRs-Fixed: 778211

Change-Id: I1ff4e2b83e627361880548a29a2b07163078d415