IMS: Don't set add participant extra if no call exist.
Dialer set add participant extra with the call intent
irrespective of call exist. This leads to no Call UI
display if dial a number immediately after add a
participant and end the conf call to use same dialpad
which launch earlier when the participant added.
Add code to set add participant extra if user clicks
on add participant and there is a call exist.

Change-Id: Ifa18b26ed9ccd368afb930382d5e8aeaf8e9511a
CRs-Fixed: 817666