Dialer: FPS is low when scrolling call logs
ListView used in call log will execute lot of SQL queries and
load lots of data, as a result, getView function will cost lots
of time to load data and draw views. This will reduce the FPS.

When list view is flinging, we try to avoid to load data and bind
views. We will reload the data for visible items in list view when
list view's scroll state is not flinging.

Change-Id: I3400f03e87a875f791abe98f8e2777e4428623a1
CRs-Fixed: 755408