Dialer: FPS is low when scrolling call logs
ListView used in call log will execute lot of SQL queries and
load lots of data, as a result, getView function will cost lots
of time to load data and draw views. This will reduce the FPS.

When list view is flinging, we try to avoid to load data and bind
views. We will reload the data for visible items in list view when
list view's scroll state is not flinging.

Change-Id: Ice476a605064cd2cd3c6a3931c7b4d18c317e4a9
CRs-Fixed: 755408